introduce more generic concept of utility to execute other dockerize tasks (word alignment, similarity checking, etc)... + some  style tests (enforced by pycodestyle/tox.ini)
