+++ nmtwizard/sampler.py
@@ -219,5 +219,6 @@ def main():
     print('res=', json.dumps(res))
     print('metadata=', json.dumps(metadata))
 
+
 if __name__ == "__main__":
     main()
