+++ tox.ini
@@ -1,0 +1,3 @@
+[pycodestyle]
+max-line-length = 120
+ignore = E402, E226, E126, E266, E302
